+++ datanode/networkhistory/service_test.go
@@ -379,12 +379,12 @@ func TestMain(t *testing.M) {
 		log.Infof("%s", goldenSourceHistorySegment[4000].HistorySegmentID)
 		log.Infof("%s", goldenSourceHistorySegment[5000].HistorySegmentID)
 
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "QmUmd4UfAxtpRQv1qpN4bmgCMbHTyMTVJF1xkx8dzB5oo4", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmQyaeEVnCzmjqZ1yQC1Qiwt374TywCK1VQKoXunBkvhbs", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmPiT7XsMpRqqMKHz4YkJrci2Hpzmv8GPQp4bGBv7M9D6a", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "QmXXdGf1U4WAdku3QvoJ9qAt7PeCGMUpdYbML4BXKkHUBn", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmbdpgppDZ3nHYktu8ukohZmkUZQzg2oBf2AUQjEpKgUUN", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "QmfEyifJ2T9zMHq1h2EPccGxe1fdDVm8rp57BKjFQ5wdN2", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "QmZVrRBX1dLuyLV32k1SvCmsrE59TtZqB6VXk7G2c8h3zq", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmSGjsHHy6V8PVJ7V44SSFo7dmEPyYFYEUpPU5VBPdqnHC", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmSPQTqAixXcijQpq56GURyX2TvAAXgEyqPkSUvR4BoJ7s", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "Qma8wkZn4PNuDwUAR9FDjGDgiq6CjBLPqBwViAfohoLdTn", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmRioFJQUpo8PB3AEN68PVe2Xq5scWmeUDRVyv3uf6VoVT", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "QmY3WQQaUrvxK363MaNhftcC6n4QLYehQaHENDH5djxMZ4", snapshots)
 	}, postgresRuntimePath, sqlFs)
 
 	if exitCode != 0 {